move the fixture to the local file
